services/consumer/requirements.txt
kafka-python==2.0.2
requests==2.32.3
boto3==1.35.24
opencv-python-headless==4.10.0.84
numpy==1.26.4
